--- gnowsys-ndf/indexb.htm
@@ -1,2 +1,0 @@
-</h1>
-<p style="Margin-top: 20px;Margin-bottom: 20px;text-align: center;" name="pg_content">
