--- gnowsys-ndf/indexc.htm
@@ -1,55 +1,0 @@
-</p>
-    </div>
-        
-            <div style="Margin-left: 20px;Margin-right: 20px;">
-      <div class="divider" style="display: block;font-size: 2px;line-height: 2px;Margin-left: auto;Margin-right: auto;width: 40px;background-color: #b4b4c4;Margin-bottom: 20px;">&nbsp;</div>
-    </div>
-        
-            <div style="Margin-left: 20px;Margin-right: 20px;">
-      <div style="line-height:5px;font-size:1px">&nbsp;</div>
-    </div>
-        
-            <div style="Margin-left: 20px;Margin-right: 20px;">
-      <div class="btn btn--flat btn--large" style="Margin-bottom: 20px;text-align: center;">
-        <![if !mso]><a style="border-radius: 4px;display: inline-block;font-size: 14px;font-weight: bold;line-height: 24px;padding: 12px 24px;text-align: center;text-decoration: none !important;transition: opacity 0.1s ease-in;color: #fff;background-color: #4eaacc;font-family: 'PT Serif', Georgia, serif;" href="http://i2c.metastudio.org/">Visit Metastudio.org</a><![endif]>
-      <!--[if mso]><p style="line-height:0;margin:0;">&nbsp;</p><v:roundrect xmlns:v="urn:schemas-microsoft-com:vml" href="http://google.com" style="width:240px" arcsize="9%" fillcolor="#4EAACC" stroke="f"><v:textbox style="mso-fit-shape-to-text:t" inset="0px,11px,0px,11px"><center style="font-size:14px;line-height:24px;color:#FFFFFF;font-family:Georgia,serif;font-weight:bold;mso-line-height-rule:exactly;mso-text-raise:4px">View Page at Metastudio.org</center></v:textbox></v:roundrect><![endif]--></div>
-    </div>
-        
-            <div style="Margin-left: 20px;Margin-right: 20px;Margin-bottom: 24px;">
-      <p style="Margin-top: 0;Margin-bottom: 0;">Powered by gstudio built at Gnowledge Lab, Homi Bhabha Centre for Science Education, TIFR.</p>
-    </div>
-        
-          </div>
-        <!--[if (mso)|(IE)]></td></tr></table><![endif]-->
-        </div>
-      </div>
-  
-      <div style="line-height:20px;font-size:20px;">&nbsp;</div>
-  
-      <div class="layout email-footer" style="Margin: 0 auto;max-width: 600px;min-width: 320px; width: 320px;width: calc(28000% - 173000px);overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;">
-        <div class="layout__inner" style="border-collapse: collapse;display: table;width: 100%;">
-        <!--[if (mso)|(IE)]><table align="center" cellpadding="0" cellspacing="0"><tr class="layout-email-footer"><td style="width: 400px;" valign="top" class="w360"><![endif]-->
-          <div class="column wide" style="text-align: left;font-size: 12px;line-height: 19px;color: #7c7e7f;font-family: Ubuntu,sans-serif;Float: left;max-width: 400px;min-width: 320px; width: 320px;width: calc(8000% - 47600px);">
-            <div style="Margin-left: 20px;Margin-right: 20px;Margin-top: 10px;Margin-bottom: 10px;">
-              
-              <div>
-                
-              </div>
-              <div style="Margin-top: 18px;">
-                <div>To update the page just reply to this mail</div>
-              </div>
-            </div>
-          </div>
-        <!--[if (mso)|(IE)]></td><td style="width: 200px;" valign="top" class="w160"><![endif]-->
-          <div class="column narrow" style="text-align: left;font-size: 12px;line-height: 19px;color: #7c7e7f;font-family: Ubuntu,sans-serif;Float: left;max-width: 320px;min-width: 200px; width: 320px;width: calc(72200px - 12000%);">
-            <div style="Margin-left: 20px;Margin-right: 20px;Margin-top: 10px;Margin-bottom: 10px;">
-              
-            </div>
-          </div>
-        <!--[if (mso)|(IE)]></td></tr></table><![endif]-->
-        </div>
-      </div>
-  </div>
-</body></html>
-
-
