--- gnowsys-ndf/report_error.py
@@ -1,9 +1,0 @@
-error_message = {"Page name required":1,
-				"Page Content required":2,
-				"User details required":3,
-				"already exists":4,
-				"User does not exist":5,
-				"User authenticated":6,
-				"User not a member of this group":7,
-				"Group does not exist":8}
-
